fix: Scroll flyout contents into bounds.
